Category deselect + create recipe  textfield improvements
Added the ability to deselect category's on the home screen which will just return the user to the home screen.

Began work on text field improvements in create recipe. If the user clicks on text field it clears it for typing, if they click off and haven't typed anything in text field returns to default text.